# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/orion-nomes-publicacaos": {
            "get": {
                "tags": [
                    "OrionNomesPublicacao"
                ],
                "summary": "getOrionNomesPublicacaoList",
                "description": "Get all OrionNomesPublicacaos",
                "operationId": "4fb7ee27b8d39537cfe933200d3285a1",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrionNomesPublicacao"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrionNomesPublicacao"
                ],
                "summary": "createOrionNomesPublicacao",
                "description": "Create OrionNomesPublicacao",
                "operationId": "89a4637e25db9eec8b52fce5c1902b1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrionNomesPublicacao"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrionNomesPublicacao"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orion-nomes-publicacaos/{id}": {
            "get": {
                "tags": [
                    "OrionNomesPublicacao"
                ],
                "summary": "getOrionNomesPublicacaoItem",
                "description": "Get OrionNomesPublicacao",
                "operationId": "fb1e198ababe1c70266d2acc525f03f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrionNomesPublicacao",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrionNomesPublicacao"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "OrionNomesPublicacao"
                ],
                "summary": "updateOrionNomesPublicacao",
                "description": "Update OrionNomesPublicacao",
                "operationId": "cf16f1f3a1078906170270cc4c6c6285",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrionNomesPublicacao",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrionNomesPublicacao"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrionNomesPublicacao"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrionNomesPublicacao"
                ],
                "summary": "deleteOrionNomesPublicacao",
                "description": "Delete OrionNomesPublicacao",
                "operationId": "ed282e1afbd4385fe67d706cddfdb36e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrionNomesPublicacao",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrionNomesPublicacao": {
                "required": [],
                "properties": {
                    "codigo": {
                        "description": "codigo",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "pf": {
                        "description": "pf",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "tipo": {
                        "description": "tipo publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "oab": {
                        "description": "oab publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "uf": {
                        "description": "uf publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "cpf": {
                        "description": "cpf publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "nome": {
                        "description": "nome publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "dj": {
                        "description": "dj publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "quem_recebe": {
                        "description": "quem recebe",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "status": {
                        "description": "status publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "possui_homonimo": {
                        "description": "Possui homônimo publicacao",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "deleted_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}